+++ src/proofs/range.rs
@@ -306,10 +306,16 @@ impl RangeDecomposition {
         ((upper_bound as f64).log2() * 3.0).ceil() as u64
     }
 
-    // We may not have floating-point arithmetics on no-std targets; thus, we use
-    // a less precise estimate.
     #[cfg(not(feature = "std"))]
     fn lower_len_estimate(upper_bound: u64) -> u64 {
+        Self::int_lower_len_estimate(upper_bound)
+    }
+
+    // We may not have floating-point arithmetics on no-std targets; thus, we use
+    // a less precise estimate.
+    #[cfg(any(test, not(feature = "std")))]
+    #[inline]
+    fn int_lower_len_estimate(upper_bound: u64) -> u64 {
         let log2_upper_bound = if upper_bound == 0 {
             0
         } else {
@@ -796,21 +802,10 @@ mod tests {
     #[test]
     #[cfg(feature = "std")]
     fn int_lower_len_estimate_is_always_not_more_than_exact() {
-        // **NB.** Must correspond to the no-std implementation of
-        // `RangeDecomposition::lower_len_estimate`.
-        fn int_lower_len_estimate(upper_bound: u64) -> u64 {
-            let log2_upper_bound = if upper_bound == 0 {
-                0
-            } else {
-                63 - u64::from(upper_bound.leading_zeros()) // rounded down
-            };
-            log2_upper_bound * 3
-        }
-
         let samples = (0..1_000).chain((1..1_000).map(|i| i * 1_000));
         for sample in samples {
             let floating_point_estimate = RangeDecomposition::lower_len_estimate(sample);
-            let int_estimate = int_lower_len_estimate(sample);
+            let int_estimate = RangeDecomposition::int_lower_len_estimate(sample);
             assert!(
                 floating_point_estimate >= int_estimate,
                 "Unexpected estimates for {}: floating-point = {}, int = {}",