+++ tests/integration/main.rs
@@ -1,4 +1,5 @@
 //! Generic tests independent of the `Group` implementation.
 
 mod basic;
+mod dkg;
 mod sharing;