+++ src/lib.rs
@@ -133,7 +133,7 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 // Documentation settings.
 #![cfg_attr(docsrs, feature(doc_cfg))]
-#![doc(html_root_url = "https://docs.rs/elastic-elgamal/0.2.1")]
+#![doc(html_root_url = "https://docs.rs/elastic-elgamal/0.3.0-beta.1")]
 // Linter settings.
 #![warn(missing_debug_implementations, missing_docs, bare_trait_objects)]
 #![warn(clippy::all, clippy::pedantic)]