+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   a Pedersen commitment in the same group. This proof can be used to switch 
   from frameworks applicable to ElGamal ciphertexts, to ones applicable to Pedersen commitments 
   (e.g., Bulletproofs for range proofs).
+- Implement distributed key generation (DKG) for threshold encryption based on Pedersen's scheme
+  with commitments of participants' keys.
 
 ### Changed