+++ CHANGELOG.md
@@ -9,19 +9,23 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 - Make `curve25519-dalek` dependency optional and do not force the choice of its math backend.
   The dependency is still enabled by default.
+
 - Allow for the `curve25519-dalek-ng` crypto backend as an alternative to `curve25519-dalek`.
   This may be beneficial for applications that use [`bulletproofs`] or other libraries 
   depending on `curve25519-dalek-ng`.
+
 - Implement zero-knowledge proof of equivalence between an ElGamal ciphertext and
   a Pedersen commitment in the same group. This proof can be used to switch 
   from frameworks applicable to ElGamal ciphertexts, to ones applicable to Pedersen commitments 
   (e.g., Bulletproofs for range proofs).
+
 - Implement distributed key generation (DKG) for threshold encryption based on Pedersen's scheme
   with commitments of participants' keys.
 
 ### Changed
 
 - Update `hashbrown` dependency.
+
 - Bump minimum supported Rust version to 1.62.
 
 ## 0.2.1 - 2022-07-04