+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 
+## 0.3.0-beta.1 - 2023-02-03
+
 ### Added
 
 - Make `curve25519-dalek` dependency optional and do not force the choice of its math backend.